create reusable Modal component
