+++ components/Modal/Modal.module.css
@@ -23,7 +23,7 @@
 }
 
 .dialog {
-  padding: 0.5rem;
+  padding: 1rem;
   margin: auto;
   display: flex;
   height: 100%;
@@ -36,41 +36,11 @@
   max-width: 500px;
   width: 100%;
   padding: 1.5rem;
+  overflow: auto;
   background-color: rgb(var(--background-end-rgb));
   border-radius: var(--border-radius);
   border: 2px solid rgb(var(--foreground-rgb));
   position: relative;
   z-index: var(--modal-z-index);
   box-shadow: 0 1px 2px 0 rgb(var(--foreground-rgb)), 0 1px 3px 1px rgb(var(--foreground-rgb));
-}
-
-.title {
-  margin-top: 0;
-}
-
-.body {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.label {
-  display: block;
-  font-weight: 600;
-  margin-bottom: 0.5rem;
-}
-
-.error {
-  color: rgb(var(--error-rgb));
-}
-
-.separator {
-  margin-top: 0.5rem;
-  margin-bottom: 0.5rem;
-  border: 2px solid rgb(var(--foreground-rgb));
-  border-radius: var(--border-radius);
-}
-
-.button {
-  font-size: 1.5rem;
 }
