+++ components/UserOnboardModal/UserOnboardModal.module.css
@@ -1,0 +1,23 @@
+.title {
+  margin-top: 0;
+}
+
+.body {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.label {
+  display: block;
+  font-weight: 600;
+  margin-bottom: 0.5rem;
+}
+
+.error {
+  color: rgb(var(--error-rgb));
+}
+
+.button {
+  font-size: 1.5rem;
+}
